Fix warning: implicit conversion changes signedness
visp/tutorial/munkres/tutorial-munkres-assignment.cpp:34:95: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   34 |     rand_ips.emplace_back(rand.uniform(10, I.getHeight() - 10), rand.uniform(10, I.getWidth() - 10));
      |                                                                      ~~~~~~~     ~~~~~~~~~~~~~^~~~
visp/tutorial/munkres/tutorial-munkres-assignment.cpp:34:58: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
   34 |     rand_ips.emplace_back(rand.uniform(10, I.getHeight() - 10), rand.uniform(10, I.getWidth() - 10));
      |                                ~~~~~~~     ~~~~~~~~~~~~~~^~~~

+++ tutorial/munkres/tutorial-munkres-assignment.cpp
@@ -31,7 +31,7 @@ int main()
   vpUniRand rand {};
   std::vector<vpImagePoint> rand_ips {};
   while (rand_ips.size() < 10) {
-    rand_ips.emplace_back(rand.uniform(10, I.getHeight() - 10), rand.uniform(10, I.getWidth() - 10));
+    rand_ips.emplace_back(rand.uniform(10, static_cast<int>(I.getHeight()) - 10), rand.uniform(10, static_cast<int>(I.getWidth()) - 10));
   }
   //! [Rand_Img_Pts]
